+++ shared-libs/ts-modules/shared/src/i18n/dictionaries/fr.ts
@@ -535,7 +535,7 @@ export default {
   579: 'doit être une clé publique SSH valide',
   580: 'Actualisation nécessaire',
   581: 'Votre interface utilisateur est mise en cache et obsolète. Essayez de recharger le PWA à l’aide du bouton ci-dessous. Si vous continuez à voir ce message, désinstallez puis réinstallez le PWA.',
-  582: 'Votre interface utilisateur est mise en cache et obsolète. Faites un rafraîchissement forcé de la page pour obtenir la dernière interface.',
+  582: 'StartOS a été mis à jour, mais cette page utilise encore l’interface précédente. Rechargez la page pour obtenir la dernière version.',
   606: 'Hôte',
   607: 'Valeur',
   610: 'DNS dynamique',
@@ -820,4 +820,5 @@ export default {
   928: 'Croissant',
   929: 'Décroissant',
   930: 'Les services de cette bibliothèque sont maintenus par la communauté Start9 et sont en phase de test bêta. Des bugs sont à prévoir. Installez-les à vos risques et périls. En cas de problème ou de question, veuillez contacter le développeur en charge du paquet.',
+  931: 'Recharger',
 } satisfies i18n